refactor(Entidades): alterar todos os comandos para spring data.

+++ rede-social/rede-social-service/src/main/java/br/com/crescer/repository/PessoaRepository.java
@@ -1,12 +1,11 @@
 package br.com.crescer.repository;
 
 import br.com.crescer.entity.Pessoa;
-import java.math.BigDecimal;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 /**
  * @author vinicius.ambrosi
  */
-public interface PessoaRepository extends PagingAndSortingRepository<Pessoa, BigDecimal> {
+public interface PessoaRepository extends PagingAndSortingRepository<Pessoa, Long> {
 
 }
