+++ rede-social/rede-social-service/src/main/java/br/com/crescer/repository/PublicacaoRepository.java
@@ -1,15 +1,12 @@
 package br.com.crescer.repository;
 
 import br.com.crescer.entity.Publicacao;
-import br.com.crescer.entity.PublicacaoConteudo;
-import java.math.BigDecimal;
-import java.util.List;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 /**
  * @author vinicius.ambrosi
  */
-public interface PublicacaoRepository extends PagingAndSortingRepository<Publicacao, BigDecimal> {
+public interface PublicacaoRepository extends PagingAndSortingRepository<Publicacao, Long> {
 
 
 }
